Добавлен класс ExtrapolationHandler
